attendance-function-app/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

# Azure Functions runtime
azure-functions

# Azure Queue SDK
azure-storage-queue

# SQL Server connection
pyodbc

# Data processing
pandas

# API calls
requests